Вынести типы сущностей из моделей
