+++ src/components/FieldsTree/FieldsTreeItem.tsx
@@ -4,7 +4,7 @@ import SearchView from 'components/SearchView';
 import TreeItem from 'components/Tree/TreeItem';
 import TreeItemLabel from 'components/Tree/TreeItemLabel';
 import isEmpty from 'helpers/isEmpty';
-import { Answer, Question, isQuestion } from 'models/fields';
+import { Answer, Question, isQuestion } from 'types/field';
 
 interface FieldsTreeItemProps {
     field: Question;
@@ -18,7 +18,7 @@ const FieldsTreeItem: React.FC<FieldsTreeItemProps> = ({ field, searchString }) 
     const render = (node: Question | Answer): JSX.Element => {
         isFound = isFound || node.searchedObject;
         uniqueKey = `${uniqueKey}-${node.id}`;
-        const nextNodeList = isQuestion(node) ? node.answerDtoList : node.openQuestonDtoList;
+        const nextNodeList = isQuestion(node) ? node.answerDtoList : node.openQuestionDtoList;
 
         return (
             <TreeItem
