+++ src/types/role.ts
@@ -1,0 +1,5 @@
+export interface Role {
+    id: number;
+    name: string;
+}
+export type RolesList = Role[];
